+++ extract_msg/__init__.py
@@ -27,8 +27,8 @@ https://github.com/TeamMsgExtractor/msg-extractor
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 __author__ = 'Destiny Peterson & Matthew Walker'
-__date__ = '2022-08-12'
-__version__ = '0.36.2'
+__date__ = '2022-08-29'
+__version__ = '0.36.3'
 
 import logging